adding states to validator
